warrior_bot/requirements.txt
alpaca-py>=0.20.0
yfinance>=0.2.37
pandas>=2.0.0
numpy>=1.24.0
schedule>=1.2.0
flask>=2.0.0
flask-socketio>=5.3.2
python-socketio>=5.8.0
